+++ src/renderer/components/dialogs/Settings-ExperimentalFeatures.tsx
@@ -1,5 +1,5 @@
-import React, { useState, useContext, FormEvent } from 'react'
-import { Card, Elevation, Radio, RadioGroup } from '@blueprintjs/core'
+import React, { useState, useContext } from 'react'
+import { Card, Elevation } from '@blueprintjs/core'
 import { RenderDTSettingSwitchType, SettingsSelector } from './Settings'
 import { ScreenContext, useTranslationFunction } from '../../contexts'
 import { DeltaInput } from '../Login-Styles'
@@ -13,6 +13,8 @@ import { DialogProps } from './DialogController'
 import SettingsStoreInstance, {
   SettingsStoreState,
 } from '../../stores/settings'
+import RadioGroup from '../RadioGroup'
+import Radio from '../Radio'
 
 const VIDEO_CHAT_INSTANCE_SYSTEMLI = 'https://meet.systemli.org/$ROOM'
 const VIDEO_CHAT_INSTANCE_AUTISTICI = 'https://vc.autistici.org/$ROOM'
@@ -132,16 +134,15 @@ export function EditVideochatInstanceDialog({
     onClose()
     onOk(configValue.trim()) // the trim is here to not save custom provider if it only contains whitespaces
   }
-  const onChangeRadio = (event: FormEvent<HTMLInputElement>) => {
-    const currentRadioValue = event.currentTarget.value as RadioButtonValue
+  const onChangeRadio = (value: string) => {
     let newConfigValue = ''
-    if (currentRadioValue === 'disabled') {
+    if (value === 'disabled') {
       newConfigValue = ''
       setRadioValue('disabled')
-    } else if (currentRadioValue === 'systemli') {
+    } else if (value === 'systemli') {
       newConfigValue = VIDEO_CHAT_INSTANCE_SYSTEMLI
       setRadioValue('systemli')
-    } else if (currentRadioValue === 'autistici') {
+    } else if (value === 'autistici') {
       newConfigValue = VIDEO_CHAT_INSTANCE_AUTISTICI
       setRadioValue('autistici')
     } else {
@@ -177,10 +178,24 @@ export function EditVideochatInstanceDialog({
             {tx('videochat_instance_explain_2')}
           </div>
 
-          <RadioGroup onChange={onChangeRadio} selectedValue={radioValue}>
+          <RadioGroup
+            onChange={onChangeRadio}
+            selectedValue={radioValue}
+            name='videochat-instance'
+          >
             <Radio key='select-none' label={tx('off')} value='disabled' />
-            <Radio key='select-systemli' label='Systemli' value='systemli' />
-            <Radio key='select-autistici' label='Autistici' value='autistici' />
+            <Radio
+              key='select-systemli'
+              label='Systemli'
+              value='systemli'
+              subtitle={VIDEO_CHAT_INSTANCE_SYSTEMLI}
+            />
+            <Radio
+              key='select-autistici'
+              label='Autistici'
+              value='autistici'
+              subtitle={VIDEO_CHAT_INSTANCE_AUTISTICI}
+            />
             <Radio
               key='select-custom'
               label={tx('custom')}